refactor(matrix): split matrix into modular subpackages and facade
split matrix logic into dedicated submodules (auth, credentials
and route through the facade
update internal calls to use facade AsyncClient and facade helpers
preserve backward compatibility by keeping matrix_utils
as the facade surface
